requirements.txt
flask
rq
requests
pymodm
telethon
pymorphy2
natasha
yargy
gunicorn
python-dotenv
dnspython
